# linkedin-insights-openapi.json
{"openapi":"3.1.0","info":{"title":"LinkedIn Insights Microservice","description":"LinkedIn Insights Microservice for scraping and analyzing LinkedIn pages","version":"1.0.0"},"paths":{"/api/v1/pages/{page_id}":{"get":{"tags":["pages"],"summary":"Get Page","description":"Get page details by LinkedIn page ID.\n\n- **page_id**: LinkedIn page ID (from URL)\n- **scrape_if_missing**: If True, scrape page if not in database","operationId":"get_page_api_v1_pages__page_id__get","parameters":[{"name":"page_id","in":"path","required":true,"schema":{"type":"string","title":"Page Id"}},{"name":"scrape_if_missing","in":"query","required":false,"schema":{"type":"boolean","default":true,"title":"Scrape If Missing"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageWithDetails"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/pages/{page_id}/scrape":{"post":{"tags":["pages"],"summary":"Scrape Page","description":"Force scrape a page and save to database.","operationId":"scrape_page_api_v1_pages__page_id__scrape_post","parameters":[{"name":"page_id","in":"path","required":true,"schema":{"type":"string","title":"Page Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageInDB"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/pages/":{"get":{"tags":["pages"],"summary":"Search Pages","description":"Search pages with filters and pagination.\n\n- **min_followers**: Minimum follower count\n- **max_followers**: Maximum follower count\n- **name**: Search by page name (partial match)\n- **industry**: Filter by industry\n- **page**: Page number\n- **size**: Items per page","operationId":"search_pages_api_v1_pages__get","parameters":[{"name":"min_followers","in":"query","required":false,"schema":{"anyOf":[{"type":"integer","minimum":0},{"type":"null"}],"title":"Min Followers"}},{"name":"max_followers","in":"query","required":false,"schema":{"anyOf":[{"type":"integer","minimum":0},{"type":"null"}],"title":"Max Followers"}},{"name":"name","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"}},{"name":"industry","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Industry"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"default":1,"title":"Page"}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":10,"title":"Size"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginatedPages"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/pages/{page_id}/employees":{"get":{"tags":["pages"],"summary":"Get Page Employees","description":"Get employees/people working at a page.\n\n- **page_id**: LinkedIn page ID\n- **limit**: Maximum number of employees to return","operationId":"get_page_employees_api_v1_pages__page_id__employees_get","parameters":[{"name":"page_id","in":"path","required":true,"schema":{"type":"string","title":"Page Id"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":20,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SocialMediaUserInDB"},"title":"Response Get Page Employees Api V1 Pages  Page Id  Employees Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/pages/{page_id}/posts":{"get":{"tags":["pages"],"summary":"Get Page Posts","description":"Get recent posts for a page.\n\n- **page_id**: LinkedIn page ID\n- **limit**: Maximum number of posts to return","operationId":"get_page_posts_api_v1_pages__page_id__posts_get","parameters":[{"name":"page_id","in":"path","required":true,"schema":{"type":"string","title":"Page Id"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":50,"minimum":1,"default":15,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PostInDB"},"title":"Response Get Page Posts Api V1 Pages  Page Id  Posts Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/pages/posts/{post_id}/comments":{"get":{"tags":["pages"],"summary":"Get Post Comments","description":"Get comments for a post.\n\n- **post_id**: Database ID of the post\n- **scrape_if_missing**: If True, scrape comments if not in database","operationId":"get_post_comments_api_v1_pages_posts__post_id__comments_get","parameters":[{"name":"post_id","in":"path","required":true,"schema":{"type":"integer","title":"Post Id"}},{"name":"scrape_if_missing","in":"query","required":false,"schema":{"type":"boolean","default":false,"title":"Scrape If Missing"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostWithComments"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/pages/{page_id}/followers-range":{"get":{"tags":["pages"],"summary":"Get Pages In Follower Range","description":"Find pages with similar follower count (±range_percent).\n\n- **page_id**: Reference page ID\n- **range_percent**: Percentage range for follower count","operationId":"get_pages_in_follower_range_api_v1_pages__page_id__followers_range_get","parameters":[{"name":"page_id","in":"path","required":true,"schema":{"type":"string","title":"Page Id"}},{"name":"range_percent","in":"query","required":false,"schema":{"type":"number","maximum":100.0,"minimum":1.0,"default":10.0,"title":"Range Percent"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","description":"Root endpoint","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health Check","description":"Health check endpoint","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"CommentInDB":{"properties":{"content":{"type":"string","title":"Content"},"commenter_name":{"type":"string","title":"Commenter Name"},"commenter_profile_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Commenter Profile Url"},"commenter_headline":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Commenter Headline"},"id":{"type":"integer","title":"Id"},"post_id":{"type":"integer","title":"Post Id"},"commented_at":{"type":"string","format":"date-time","title":"Commented At"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["content","commenter_name","id","post_id","commented_at","created_at"],"title":"CommentInDB"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"PageInDB":{"properties":{"page_id":{"type":"string","title":"Page Id"},"name":{"type":"string","title":"Name"},"url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Url"},"linkedin_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Linkedin Id"},"profile_picture_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Profile Picture Url"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"website":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Website"},"industry":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Industry"},"total_followers":{"type":"integer","title":"Total Followers","default":0},"head_count":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Head Count"},"specialities":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Specialities"},"location":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Location"},"founded_year":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Founded Year"},"company_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Company Type"},"id":{"type":"integer","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated At"},"last_scraped_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Last Scraped At"}},"type":"object","required":["page_id","name","id","created_at"],"title":"PageInDB"},"PageWithDetails":{"properties":{"page_id":{"type":"string","title":"Page Id"},"name":{"type":"string","title":"Name"},"url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Url"},"linkedin_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Linkedin Id"},"profile_picture_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Profile Picture Url"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"website":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Website"},"industry":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Industry"},"total_followers":{"type":"integer","title":"Total Followers","default":0},"head_count":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Head Count"},"specialities":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Specialities"},"location":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Location"},"founded_year":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Founded Year"},"company_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Company Type"},"id":{"type":"integer","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated At"},"last_scraped_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Last Scraped At"},"posts_count":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Posts Count"},"employees_count":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Employees Count"}},"type":"object","required":["page_id","name","id","created_at"],"title":"PageWithDetails"},"PaginatedPages":{"properties":{"items":{"items":{"$ref":"#/components/schemas/PageInDB"},"type":"array","title":"Items"},"total":{"type":"integer","title":"Total"},"page":{"type":"integer","title":"Page"},"size":{"type":"integer","title":"Size"},"pages":{"type":"integer","title":"Pages"}},"type":"object","required":["items","total","page","size","pages"],"title":"PaginatedPages"},"PostInDB":{"properties":{"linkedin_post_id":{"type":"string","title":"Linkedin Post Id"},"content":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Content"},"image_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Image Url"},"video_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Video Url"},"post_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Post Url"},"likes_count":{"type":"integer","title":"Likes Count","default":0},"comments_count":{"type":"integer","title":"Comments Count","default":0},"shares_count":{"type":"integer","title":"Shares Count","default":0},"id":{"type":"integer","title":"Id"},"page_id":{"type":"integer","title":"Page Id"},"posted_at":{"type":"string","format":"date-time","title":"Posted At"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated At"}},"type":"object","required":["linkedin_post_id","id","page_id","posted_at","created_at"],"title":"PostInDB"},"PostWithComments":{"properties":{"linkedin_post_id":{"type":"string","title":"Linkedin Post Id"},"content":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Content"},"image_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Image Url"},"video_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Video Url"},"post_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Post Url"},"likes_count":{"type":"integer","title":"Likes Count","default":0},"comments_count":{"type":"integer","title":"Comments Count","default":0},"shares_count":{"type":"integer","title":"Shares Count","default":0},"id":{"type":"integer","title":"Id"},"page_id":{"type":"integer","title":"Page Id"},"posted_at":{"type":"string","format":"date-time","title":"Posted At"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated At"},"comments":{"items":{"$ref":"#/components/schemas/CommentInDB"},"type":"array","title":"Comments","default":[]}},"type":"object","required":["linkedin_post_id","id","page_id","posted_at","created_at"],"title":"PostWithComments"},"SocialMediaUserInDB":{"properties":{"linkedin_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Linkedin Id"},"name":{"type":"string","title":"Name"},"profile_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Profile Url"},"profile_picture_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Profile Picture Url"},"headline":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Headline"},"current_position":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Current Position"},"id":{"type":"integer","title":"Id"},"page_id":{"type":"integer","title":"Page Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated At"}},"type":"object","required":["name","id","page_id","created_at"],"title":"SocialMediaUserInDB"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}